+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -3,8 +3,23 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.175.0
+------------------------------
+*August 25, 2021*
+
+### Removed
+- Age verification div wrapper from `AgeVerification.vue` component.
+
+### Changed
+- `AgeVerificationError` ref location from wrapper to `errorMessage` component.
+- `handleAgeVerifcation` method to look for correct component when error appears.
+
+### Fixed
+- Typo in `AgeVerification.vue` component.
+
 
 v0.174.0
+------------------------------
 *August 25, 2021*
 
 ### Added